添加GetSelf接口
其它修复项：
1. 修复了注册接口没有正确放行的问题
2. 学生端Controller添加了相关注解用于Swagger调试

+++ teacher/src/main/java/org/nuist/business_object/KnowledgeBO.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class KnowledgeBO {
-    private int knowledgeId;
+    private Long knowledgeId;
     private String name;
     private String description;
     private String difficultyLevel;
